fix: macOS __int128 operator<< error in IncrementalTrainer
On macOS (libc++), std::filesystem::file_time_type::duration::rep is
__int128, which has no operator<< overload. This caused an 'ambiguous
overloaded operator' compilation error at:
  IncrementalTrainer.cpp:1130 — compute_data_checksum()

Fix: cast time_since_epoch().count() to long long before streaming.
The cast is safe for any plausible file modification time.

+++ src/IncrementalTrainer.cpp
@@ -1127,7 +1127,9 @@ std::string IncrementalTrainer::compute_data_checksum(const std::string& data_fi
     auto ftime = fs::last_write_time(data_file);
 
     std::ostringstream oss;
-    oss << size << "_" << ftime.time_since_epoch().count();
+    // file_time_type::duration::rep is __int128 on macOS (libc++) which has no
+    // operator<< overload — cast to long long to keep it portable.
+    oss << size << "_" << static_cast<long long>(ftime.time_since_epoch().count());
 
     return oss.str();
 }
